+++ libraries/transformer/src/main/java/androidx/media3/transformer/EditedMediaItem.java
@@ -15,21 +15,29 @@
  */
 package androidx.media3.transformer;
 
+import static androidx.media3.transformer.TransformerUtil.containsSpeedChangingEffects;
+import static androidx.media3.transformer.TransformerUtil.validateSpeedChangingEffects;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.Preconditions.checkState;
 import static java.lang.Math.max;
 
 import androidx.annotation.IntRange;
+import androidx.annotation.Nullable;
 import androidx.media3.common.C;
 import androidx.media3.common.Effect;
 import androidx.media3.common.MediaItem;
 import androidx.media3.common.audio.AudioProcessor;
+import androidx.media3.common.audio.SpeedChangingAudioProcessor;
+import androidx.media3.common.audio.SpeedProvider;
 import androidx.media3.common.util.Log;
+import androidx.media3.common.util.SpeedProviderUtil;
 import androidx.media3.common.util.UnstableApi;
 import androidx.media3.common.util.Util;
+import androidx.media3.effect.TimestampAdjustment;
 import androidx.media3.exoplayer.source.MediaSource;
 import androidx.media3.extractor.mp4.Mp4Extractor;
+import com.google.common.collect.ImmutableList;
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
 import java.util.Objects;
 import org.json.JSONException;
@@ -50,6 +58,8 @@ public final class EditedMediaItem {
     private long durationUs;
     private int frameRate;
     private Effects effects;
+    private SpeedProvider speedProvider;
+    private boolean allowMatchingSpeedChangingEffectForSpeedProvider;
 
     /**
      * Creates an instance.
@@ -75,6 +85,7 @@ public final class EditedMediaItem {
               : Util.msToUs(mediaItem.localConfiguration.imageDurationMs);
       frameRate = C.RATE_UNSET_INT;
       effects = Effects.EMPTY;
+      speedProvider = SpeedProvider.DEFAULT;
     }
 
     private Builder(EditedMediaItem editedMediaItem) {
@@ -85,6 +96,9 @@ public final class EditedMediaItem {
       this.durationUs = editedMediaItem.durationUs;
       this.frameRate = editedMediaItem.frameRate;
       this.effects = editedMediaItem.effects;
+      this.speedProvider = editedMediaItem.speedProvider;
+      this.allowMatchingSpeedChangingEffectForSpeedProvider =
+          editedMediaItem.allowMatchingSpeedChangingEffectForSpeedProvider;
     }
 
     /**
@@ -219,6 +233,23 @@ public final class EditedMediaItem {
       return this;
     }
 
+    /**
+     * Sets a {@link SpeedProvider} to control the presentation speed of the {@link
+     * EditedMediaItem}.
+     *
+     * <p>If a {@link SpeedProvider} is set, speed changing {@linkplain #setEffects effects} are not
+     * allowed.
+     *
+     * <p>The default value is {@link SpeedProvider#DEFAULT}, which represents an unmodified speed.
+     *
+     * @return This builder.
+     */
+    @CanIgnoreReturnValue
+    public Builder setSpeed(SpeedProvider provider) {
+      this.speedProvider = provider;
+      return this;
+    }
+
     /** Builds an {@link EditedMediaItem} instance. */
     public EditedMediaItem build() {
       return new EditedMediaItem(this);
@@ -233,6 +264,36 @@ public final class EditedMediaItem {
     @CanIgnoreReturnValue
     /* package */ Builder setMediaItem(MediaItem mediaItem) {
       this.mediaItem = mediaItem;
+      return this;
+    }
+
+    /**
+     * Sets the provided {@link SpeedChangingAudioProcessor} and {@link TimestampAdjustment} as the
+     * first elements of their respective pipelines.
+     *
+     * <p>The effects' {@link SpeedProvider} must match the one set by {@link
+     * #setSpeed(SpeedProvider)}.
+     *
+     * @return This builder.
+     */
+    @CanIgnoreReturnValue
+    /* package */ Builder setSpeedChangingEffects(
+        SpeedChangingAudioProcessor processor, @Nullable TimestampAdjustment effect) {
+      checkArgument(effect == null || processor.getSpeedProvider() == effect.speedProvider);
+      this.allowMatchingSpeedChangingEffectForSpeedProvider = true;
+      ImmutableList<AudioProcessor> audioProcessors =
+          new ImmutableList.Builder<AudioProcessor>()
+              .add(processor)
+              .addAll(effects.audioProcessors)
+              .build();
+      ImmutableList<Effect> videoEffects =
+          effect == null
+              ? effects.videoEffects
+              : new ImmutableList.Builder<Effect>()
+                  .add(effect)
+                  .addAll(effects.videoEffects)
+                  .build();
+      this.effects = new Effects(audioProcessors, videoEffects);
       return this;
     }
   }
@@ -279,6 +340,10 @@ public final class EditedMediaItem {
   /** The {@link Effects} to apply to the {@link #mediaItem}. */
   public final Effects effects;
 
+  public final SpeedProvider speedProvider;
+
+  private final boolean allowMatchingSpeedChangingEffectForSpeedProvider;
+
   /** The duration for which this {@code EditedMediaItem} should be presented, in microseconds. */
   private long presentationDurationUs;
 
@@ -287,11 +352,21 @@ public final class EditedMediaItem {
         !builder.removeAudio || !builder.removeVideo, "Audio and video cannot both be removed");
     if (isGap(builder.mediaItem)) {
       checkArgument(builder.durationUs != C.TIME_UNSET);
-      checkArgument(
-          !builder.removeAudio
-              && !builder.flattenForSlowMotion
-              && builder.effects.audioProcessors.isEmpty());
+      checkArgument(!builder.removeAudio);
+      checkArgument(!builder.flattenForSlowMotion);
+      checkArgument(builder.effects.audioProcessors.isEmpty());
+      checkArgument(builder.speedProvider == SpeedProvider.DEFAULT);
     }
+
+    if (builder.speedProvider != SpeedProvider.DEFAULT) {
+      if (builder.allowMatchingSpeedChangingEffectForSpeedProvider) {
+        checkState(validateSpeedChangingEffects(builder.effects, builder.speedProvider));
+        checkState(!containsSpeedChangingEffects(builder.effects, /* ignoreFirstEffect= */ true));
+      } else {
+        checkState(!containsSpeedChangingEffects(builder.effects, /* ignoreFirstEffect= */ false));
+      }
+    }
+
     this.mediaItem = builder.mediaItem;
     this.removeAudio = builder.removeAudio;
     this.removeVideo = builder.removeVideo;
@@ -299,6 +374,9 @@ public final class EditedMediaItem {
     this.durationUs = builder.durationUs;
     this.frameRate = builder.frameRate;
     this.effects = builder.effects;
+    this.speedProvider = builder.speedProvider;
+    this.allowMatchingSpeedChangingEffectForSpeedProvider =
+        builder.allowMatchingSpeedChangingEffectForSpeedProvider;
     presentationDurationUs = C.TIME_UNSET;
   }
 
@@ -360,6 +438,10 @@ public final class EditedMediaItem {
    * @param durationUs The input duration in microseconds.
    */
   /* package */ long getDurationAfterEffectsApplied(long durationUs) {
+    if (speedProvider != SpeedProvider.DEFAULT) {
+      return SpeedProviderUtil.getDurationAfterSpeedProviderApplied(speedProvider, durationUs);
+    }
+
     long audioDurationUs = durationUs;
     long videoDurationUs = durationUs;
     if (removeAudio) {