+++ libraries/transformer/src/test/java/androidx/media3/transformer/EditedMediaItemBuilderTest.java
@@ -26,6 +26,7 @@ import androidx.media3.common.MediaItem;
 import androidx.media3.common.audio.AudioProcessor;
 import androidx.media3.common.audio.SpeedChangingAudioProcessor;
 import androidx.media3.common.audio.SpeedProvider;
+import androidx.media3.effect.TimestampAdjustment;
 import androidx.media3.test.utils.TestSpeedProvider;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import com.google.common.collect.ImmutableList;
@@ -200,6 +201,77 @@ public final class EditedMediaItemBuilderTest {
             .build();
 
     assertThat(editedMediaItem.getPresentationDurationUs()).isEqualTo(500_000);
+  }
+
+  @Test
+  public void setSpeed_withSpeedChangingAudioProcessor_throws() {
+    SpeedChangingAudioProcessor processor = new SpeedChangingAudioProcessor(SpeedProvider.DEFAULT);
+    Effects effects = new Effects(ImmutableList.of(processor), ImmutableList.of());
+    assertThrows(
+        IllegalStateException.class,
+        () ->
+            new EditedMediaItem.Builder(MediaItem.EMPTY)
+                .setSpeed(SPEED_PROVIDER_2X)
+                .setEffects(effects)
+                .build());
+  }
+
+  @Test
+  public void setSpeed_withTimestampAdjustment_throws() {
+    TimestampAdjustment timestampAdjustment =
+        new TimestampAdjustment((inputTimeUs, outputTimeConsumer) -> {}, SpeedProvider.DEFAULT);
+    Effects effects = new Effects(ImmutableList.of(), ImmutableList.of(timestampAdjustment));
+    assertThrows(
+        IllegalStateException.class,
+        () ->
+            new EditedMediaItem.Builder(MediaItem.EMPTY)
+                .setSpeed(SPEED_PROVIDER_2X)
+                .setEffects(effects)
+                .build());
+  }
+
+  @Test
+  public void setSpeed_withSetSpeedChangingEffects_doesNotThrow() {
+    TimestampAdjustment timestampAdjustment =
+        new TimestampAdjustment((inputTimeUs, outputTimeConsumer) -> {}, SPEED_PROVIDER_2X);
+    SpeedChangingAudioProcessor processor = new SpeedChangingAudioProcessor(SPEED_PROVIDER_2X);
+    EditedMediaItem unused =
+        new EditedMediaItem.Builder(MediaItem.EMPTY)
+            .setSpeed(SPEED_PROVIDER_2X)
+            .setSpeedChangingEffects(processor, timestampAdjustment)
+            .build();
+  }
+
+  @Test
+  public void setSpeed_withSetSpeedChangingEffects_throwsWithMismatchingSpeedProviders() {
+    SpeedChangingAudioProcessor processor = new SpeedChangingAudioProcessor(SpeedProvider.DEFAULT);
+    assertThrows(
+        IllegalStateException.class,
+        () ->
+            new EditedMediaItem.Builder(MediaItem.EMPTY)
+                .setSpeed(SPEED_PROVIDER_2X)
+                .setSpeedChangingEffects(processor, /* effect= */ null)
+                .build());
+  }
+
+  @Test
+  public void setSpeedChangingEffects_withMismatchingSpeedProviders_throws() {
+    TimestampAdjustment timestampAdjustment =
+        new TimestampAdjustment((inputTimeUs, outputTimeConsumer) -> {}, SPEED_PROVIDER_2X);
+    SpeedChangingAudioProcessor processor = new SpeedChangingAudioProcessor(SpeedProvider.DEFAULT);
+    assertThrows(
+        IllegalArgumentException.class,
+        () ->
+            new EditedMediaItem.Builder(MediaItem.EMPTY)
+                .setSpeedChangingEffects(processor, timestampAdjustment)
+                .build());
+  }
+
+  @Test
+  public void getDurationAfterEffectsApplied_withSpeedProvider_returnsCorrectDuration() {
+    EditedMediaItem editedMediaItem =
+        new EditedMediaItem.Builder(MediaItem.EMPTY).setSpeed(SPEED_PROVIDER_2X).build();
+    assertThat(editedMediaItem.getDurationAfterEffectsApplied(1_000_000)).isEqualTo(500_000);
   }
 
   @Test